requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.4.0
Backtesting==0.3.3
bokeh==2.4.3
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cloud-sql-python-connector==0.6.2
cryptography==37.0.2
cycler==0.11.0
dash==2.5.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.1.2
Flask-Compress==1.12
fonttools==4.33.3
frozenlist==1.3.0
google-auth==2.9.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
multidict==6.0.2
numpy==1.23.0
packaging==21.3
pandas==1.4.3
pg8000==1.29.1
Pillow==9.2.0
plotly==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rsa==4.8
scramp==1.4.1
six==1.16.0
SQLAlchemy==1.4.39
tenacity==8.0.1
tornado==6.2
typing-extensions==4.3.0
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
